requirements.txt
azure-core==1.25.1
azure-identity==1.10.0
azure-mgmt-resource==21.1.0
